+++ src/main/java/com/atomic/repositories/AtomicRepository.java
@@ -3,5 +3,5 @@ package com.atomic.repositories;
 import com.atomic.models.Atomic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtomicRepository extends JpaRepository<Atomic, Long> {
+public interface AtomicRepository extends JpaRepository<Atomic, Integer> {
 }
